Downloads, extracts and configures hadoop.
Changed name for download-hadoop to 
hadoop-download and the target uses properties
set up for hadoop. It's also made sure that the 
download only happens if it's not already 
downloaded.

Extraction is only made if hadoop is downloaded. 
This removes temporary coupling between the 
download target and the extract target.

The configuration of hadoop is made by copying
pre made conf files which live in 
contrib/hadoop-utils/conf. They contain values 
which are replaced by properties and environment 
variables. By doing this, we can configure our 
local hadoop instance any way we would want to.

It's now possible for the user to set which 
hadoop_home and/or splunk_home they want to run the
tests on. hadoop-env and hadoop-setup are removed
with this change. Targets which depended on
hadoop-env or hadoop-setup targets has been 
adjusted accordingly. 

Also added splunk-unpack and splunk-clean-all 
which are needed for the build.xml to keep the 
file in a stable state.

This commit is big because I thought I could split
my work into smaller commits with git add -i. 
It is possible to split the commits, but not as 
much as I wanted to.