Path resolver can create fully qualified URIs for the used FileSystem. So both local and hdfs file systems can be used.
The PathResolver now takes a WritableFileSystem for
being able to get the URI schema, authority and
a writable path on the FileSystem. Without it
it can't create a fully qualified URI that can
just be used by the ArchivingFileSystem.

In addition, the BucketArchiver had to change the
way it created the PathResolver. And so the
construction code was moved to a BucketArchiverFactory,
and the ArchiverFunctionalTest had to be changed.