BucketFreezer takes index name as an argument. Major change in Bucket utils. Bucket doesn't have requirements on the path that a bucket is on.
The index name is specified to the BucketFreezer as an argument. The bucket frezer then moves the bucket to a safe location which is "safe/location/root/INDEX_NAME/BUCKET_NAME".

The changes on Bucket utils is to make it more dynamic. There are also changes on File utils to do some clean up after it self (deleteOnExit).

The Bucket class now gets the index name as an argument as doesn't depend on a file path structure.