Separated tests and non-tests build directory to separate jars in a scalable way.
Before the separation of tests and non-tests we
had to include and exclude test classes manually
in build.xml, to separate our classes in our jars.
This also means that every time someone adds a new
package, they must know of this and manually edit
the build.xml.
Now classes and packages can be added without having
to do anything else, for everything to work.