Fixed a bug introduced when minimizing code duplication in build.xml.
Ant doesn't work the expected way. A property (condition property) is immutable; will not be evaluated again and therefore hold the same value as before.

In this case The first check found that Hadoop wasn't downloaded and downloaded it and the second check also found that Hadoop weren't downloaded and therefore did not extract Hadoop