HADOOP-172 - Changed archiver logging to use log4j but still use the implemented archiver-logging-format (for example "did", "happen", "expected").
ArchiverLogger is now LogFormatter.
Why? Because log4j couldn't recognize the actual method calling the logging method when ArchiverLogger handled the communication to log4j. Every class with log-messages now uses its own logger to help log4j detect the method that logs (This is an optional output format option, there is still no excuse to exclude it though).

Changed the test to check the archiver-logging-format.