UtilsMBean uses a Test MBean for real configuration.
When testing rest calls against the Jetty server
started by Splunk, and also using code in Eclipse
it's possible to get the same configuration by
using the ShepArchiverForTests class. It just has
a different path to the archiver.xml.
This way of getting the configuration should probably
be changed though.