Changed where bucket.size is stored. #13
To get all functional tests to run, the
location where bucket.size was stored needed
to be changed. It's now stored in an archived
bucket, in a new directory called archive_meta.
The bucket size feature now depends on the full
bucket archive path, to be able to read the
bucket.size file.

Also, should the archive_meta directory be
removed when the bucket is thawed back?