Can archive multiple formats simultaneous. #48
Huge commit, sorry, but it does the job.

Changed the configuration so it takes
a list for bucket formats for archiving
instead of a single format. This change
makes it possible to archive more than
one format of a bucket in a single run.
Changes to the configuration interface
requires a lot of other changes.