Made ArchiveConfiguration safer. #18
It was possible to configure the archiver
to have the path for temp data and
archiver data, to a value that was not
safe to delete. Instead, the paths
for data to be deleted is created inside
a folder inside the configured value,
making it safer to delete all the data
that is created by the archiver.