Config values are set at runtime by build.xml.
Revised the configuration implementation
again. Now we have more control over what
values are set to the mbeans. It's not
staticly written values as before, but
dynamically set values by the build.xml
that can depend on the environment.
The classes that use configuration can
choose between using the configurations
set by the build.xml, which also happens
to be the same values as the running
Shuttl is using, or they can use configs
that has all values set to null OR they
can use a file which they choose them
selves.
Which configuration to use is much more
configurable now.