Extracted a very unecessary dependency on java.util.URI. #74.
This is probably the bigest change ever
refactored on the system. The URI class
was nested deep into the system for no
real reason. It was assumed that all
backends would have a URI scheme, and
that we keep the URI at all times on
all buckets. This was not the case, and
everything from the Bucket model to the
PathResolver had to be refactored.

The change was made in the first place
because glacier does not have a URI
that we can just call. Instead we want
to configure each backend with key-values
instead of with URI's.

This change to the system should also
make it easier to transition into
a key-value store, since the abstraction
for transfering buckets is, 'give this
bucket this key, and when I give you
the key again then give me back my bucket'.