Glacier thawing actually works. Now with tests. #105
GlacierClient's download abstraction
is now based on the single file that
should be thawed, instead of having
to figure out the original file name
of the file. GlacierArchiveFileSystem
uses this new abstraction properly.

Note: BucketFormat might not be suitable
as an Enum for much longer, since all
formats shouldn't have a known extension.