Configuration tries to resolve ShuttlHome
Configuration paths are gotten relative to
Shuttl's installation paths.
There's currently no Splunk REST endpoint
to get the applications directory, so first
we try to get the shuttl-<version>.jar's
path, and then from there resolve shuttl
home. If that doesn't work, and to be
backwards compatible with tests and stuff,
we try $SPLUNK_HOME/etc/apps/shuttl.

The reason we go for $SPLUNK_HOME second
is because we want search head pooling
to be supported, and it'll work if we
locate shuttl home from the jar.